Addressed review comments:
- don't exit in cluster-up check in generate.sh
- fail operator start with invalid config from env vars
- set PUSH_LOG_FILE per default

+++ hack/generate.sh
@@ -69,28 +69,29 @@ mockgen -source pkg/handler-launcher-com/cmd/info/info.pb.go -package=info -dest
 # update cluster-up if needed
 version_file="cluster-up/version.txt"
 sha_file="cluster-up-sha.txt"
+download_cluster_up=true
 function getClusterUpShasum() {
     find ${KUBEVIRT_DIR}/cluster-up -type f | sort | xargs sha1sum | sha1sum | awk '{print $1}'
 }
-
 # check if we got a new cluster-up git commit hash
 if [[ -f "${version_file}" ]] && [[ $(cat ${version_file}) == ${kubevirtci_git_hash} ]]; then
     # check if files are modified
     current_sha=$(getClusterUpShasum)
     if [[ -f "${sha_file}" ]] && [[ $(cat ${sha_file}) == ${current_sha} ]]; then
         echo "cluster-up is up to date and not modified"
-        exit
+        download_cluster_up=false
     else
-        echo "cluster up was modified"
+        echo "cluster-up was modified"
     fi
 else
     echo "cluster-up git commit hash was updated"
 fi
+if [[ "$download_cluster_up" == true ]]; then
+    echo "downloading cluster-up"
+    rm -rf cluster-up
+    curl -L https://github.com/kubevirt/kubevirtci/archive/${kubevirtci_git_hash}/kubevirtci.tar.gz | tar xz kubevirtci-${kubevirtci_git_hash}/cluster-up --strip-component 1
 
-echo "downloading cluster-up"
-rm -rf cluster-up
-curl -L https://github.com/kubevirt/kubevirtci/archive/${kubevirtci_git_hash}/kubevirtci.tar.gz | tar xz kubevirtci-${kubevirtci_git_hash}/cluster-up --strip-component 1
-
-echo ${kubevirtci_git_hash} >${version_file}
-new_sha=$(getClusterUpShasum)
-echo ${new_sha} >${sha_file}
+    echo ${kubevirtci_git_hash} >${version_file}
+    new_sha=$(getClusterUpShasum)
+    echo ${new_sha} >${sha_file}
+fi
